third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-state-changes-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests MediaSource API state transitions.

webkitSourceState : closed
EVENT(webkitsourceopen) : open
onFirstSourceOpen
EVENT(webkitsourceended) : ended
onFirstSourceEnded
triggerFirstSeek
EVENT(seeking)
onFirstSeeking
EVENT(webkitsourceopen) : open
EVENT(seeked)
onFirstSeeked
EVENT(seeking)
onSecondSeeking
EVENT(seeked)
onSecondSeeked
EVENT(webkitsourceclose) : closed
onFirstSourceClose
EVENT(emptied)
EVENT(webkitsourceopen) : open
onSecondSourceOpen
EVENT(webkitsourceended) : ended
onSecondSourceEnded
triggerSecondSourceClose
EVENT(webkitsourceclose) : closed
onSecondSourceClose
EVENT(emptied)
EVENT(webkitsourceopen) : open
onThirdSourceOpen
END OF TEST
